max components changed to 8

+++ src/ecs/mod.rs
@@ -3,5 +3,5 @@ pub mod entity;
 pub mod system;
 pub mod systems;
 
-pub type Signature = u16;
+pub type Signature = u8;
 pub type ComponentId = u8;
